remove wf spec from ds when deleting workflow
