+++ deployment/data-feeds/changeset/jd_propose_wf_jobs_v2.go
@@ -23,10 +23,6 @@ const (
 	timeoutV2 = 240 * time.Second
 )
 
-type WorkflowMetadata struct {
-	Workflows map[string]string
-}
-
 // ProposeWFJobsToJDV2Changeset is a Durable Pipeline compatible changeset that reads a feed state file,
 // creates a workflow job spec from it and proposes it to JD.
 var ProposeWFJobsToJDV2Changeset = cldf.CreateChangeSet(proposeWFJobsToJDV2Logic, proposeWFJobsToJDV2Precondition)
@@ -88,7 +84,7 @@ func proposeWFJobsToJDV2Logic(env cldf.Environment, c types.ProposeWFJobsV2Confi
 		return cldf.ChangesetOutput{}, fmt.Errorf("failed to create workflow spec: %w", err)
 	}
 
-	// log the workflow spec for debugging purposes. We don't yet store it anywhere
+	// log the workflow spec for debugging purposes.
 	fmt.Println(workflowSpec)
 
 	// create workflow job spec TOML
@@ -107,29 +103,7 @@ func proposeWFJobsToJDV2Logic(env cldf.Environment, c types.ProposeWFJobsV2Confi
 	// Save workflow spec in the datastore
 	ds := datastore.NewMemoryDataStore()
 
-	// environment metadata is overwritten with every Set(), so we need to read the existing metadata first
-	record, err := env.DataStore.EnvMetadata().Get()
-	if err != nil {
-		env.Logger.Errorf("failed to get env datastore: %s", err)
-	}
-
-	metadata, err := datastore.As[WorkflowMetadata](record.Metadata)
-	if err != nil {
-		env.Logger.Errorf("failed to cast env metadata: %s", err)
-	}
-
-	if metadata.Workflows == nil {
-		metadata.Workflows = make(map[string]string)
-	}
-
-	// upsert the workflow spec in the metadata
-	metadata.Workflows[workflowSpecConfig.WorkflowName] = workflowSpec
-
-	err = ds.EnvMetadata().Set(
-		datastore.EnvMetadata{
-			Metadata: metadata,
-		},
-	)
+	err = UpdateWorkflowMetadataDS(env, ds, workflowSpecConfig.WorkflowName, workflowSpec)
 	if err != nil {
 		env.Logger.Errorf("failed to set workflow spec in datastore: %s", err)
 	}
